+++ PawNClaw.Backend/PawNClaw.Data/Parameter/CreateAdminParameter.cs
@@ -11,10 +11,8 @@ namespace PawNClaw.Data.Parameter
         public string UserName { get; set; }
         public int CreatedUser { get; set; }
         public string Phone { get; set; }
-
-        public byte Gender { get; set; }
+        public int Gender { get; set; }
         public readonly string RoleCode = "MOD";
-
         public string Name { get; set; }
     }
 }
